Refactor cpfp & single-block gbt code into mini-miner module
